Towers
Made three different types of towers:
1. Basic - Decent Range, Decent DMG, Decent Attack Speed.
2. Fast - Lower Range, Lower Damage, High Attack Speed.
3. Slow Type - Decent Range, Low DMG, Slows enemies.

+++ Assets/Scripts/HealthBarBT.cs
@@ -17,13 +17,13 @@ public class HealthBarBT : MonoBehaviour
     public Slider btSlider;
     public Image fill;
 
-    public void SetMaxHealthBT(int health)
+    public void SetMaxHealthBT(float health)
     {
         btSlider.maxValue = health;
         btSlider.value = health;
     }
 
-    public void SetHealthBT(int health)
+    public void SetHealthBT(float health)
     {
         btSlider.value = health;
     }
